quantum_modelling/requirements.txt
numpy
pandas
yfinance
matplotlib
scipy
dwave-ocean-sdk
